[software] Add sqrt header file
