+++ software/apps/choldec_schedule/mempool_cholesky_q32s_schedule.h
@@ -4,6 +4,8 @@
 
 // Author: Marco Bertuletti, ETH Zurich
 
+#include "mempool_sqrt_q32s.h"
+
 int32_t mempool_sqrt_q32s(int32_t Src, const uint32_t fracBits);
 
 void mempool_cholesky_q32s(int32_t *pSrc, int32_t *pL, const uint32_t n,
